requirements.txt
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
pytz==2024.1
psycopg2-binary==2.9.9
